requirements.txt
flask
python-telegram-bot
anthropic
python-dotenv
google-generativeai
anthropic 